+++ week8/PhotoInspiration/app/src/main/java/com/manuflowers/photoinspiration/worker/SynchronizeServiceWorker.kt
@@ -1,14 +1,8 @@
 package com.manuflowers.photoinspiration.worker
 
-import android.app.NotificationChannel
-import android.app.NotificationManager
 import android.content.Context
-import android.content.Context.NOTIFICATION_SERVICE
-import android.os.Build
-import androidx.core.app.NotificationCompat
 import androidx.work.CoroutineWorker
 import androidx.work.WorkerParameters
-import com.manuflowers.photoinspiration.R
 import com.manuflowers.photoinspiration.application.PhotoInspirationApplication
 import com.manuflowers.photoinspiration.data.PhotosInspirationRepository
 import com.manuflowers.photoinspiration.data.local.database.PhotoInspirationDatabase
@@ -16,6 +10,7 @@ import com.manuflowers.photoinspiration.data.local.database.PhotosDao
 import com.manuflowers.photoinspiration.data.local.preferences.PhotoInspirationPreferences
 import com.manuflowers.photoinspiration.data.remote.networking.RemoteApiManager
 import com.manuflowers.photoinspiration.data.remote.networking.buildApiService
+import com.manuflowers.photoinspiration.util.showNotification
 
 const val NOTIFICATION_CHANNEL_NAME = "Synchronize service channel"
 const val NOTIFICATION_CHANNEL_ID = "Synchronize ID"
@@ -42,38 +37,11 @@ class SynchronizeServiceWorker(context: Context, workerParameters: WorkerParamet
     override suspend fun doWork(): Result {
         return try {
             repository.fetchAndSavePhotos()
-            showNotification()
+            showNotification(applicationContext)
             Result.success()
         } catch (error: Throwable) {
             error.printStackTrace()
             Result.failure()
-        }
-    }
-
-    private fun showNotification() {
-        val notificationManager = (applicationContext.getSystemService(NOTIFICATION_SERVICE)) as NotificationManager
-
-        createNotificationChannel(notificationManager)
-
-        val builder = NotificationCompat.Builder(applicationContext, NOTIFICATION_CHANNEL_ID)
-            .setSmallIcon(R.drawable.ic_app)
-            .setContentTitle("Synchronization service")
-            .setContentText("Synchronizing home")
-            .setPriority(NotificationCompat.PRIORITY_DEFAULT)
-            .build()
-
-        notificationManager.notify(1, builder)
-
-    }
-
-    private fun createNotificationChannel(manager: NotificationManager?) {
-        if(Build.VERSION.SDK_INT>= Build.VERSION_CODES.O) {
-            val serviceChannel = NotificationChannel(
-                    NOTIFICATION_CHANNEL_ID,
-            NOTIFICATION_CHANNEL_NAME,
-            NotificationManager.IMPORTANCE_DEFAULT
-            )
-            manager?.createNotificationChannel(serviceChannel)
         }
     }
